requirements.txt
numpy==1.26.4
pandas
matplotlib
scikit-learn
seaborn
wordcloud
environs
fake-useragent==1.5.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
openpyxl==3.1.3
apache-airflow==2.9.1